requirements.txt
Flask>=0.9
python-memcache==1.48
pysaml2==0.4.2
